Add Operations tab to Tag page
Add a new Operations tab to the Tags page and move the Auto tag button
to this tab. This will allow for adding more operations, for now being
merging of tags.
